bitkeeper revision 1.879.1.5 (40912106hCc_Qr0pmIBFCRUoKbpCcQ)
Add support for building domains with IO privileges.

With a non-empty pci_device_list, xc_dom_create.py will automatically
cause the SIF_PRIVILEGED flag to be set in the domain's start info
structure.  This notifies it to do an IOPL, in turn allowing it to use
its PCI devices.