bitkeeper revision 1.1098.1.1 (4107c16f_jhuAGUiMBcotuiFhO6dHQ)
Add a simple debugger for Xen and domain 0.  This isn't anything
like as complete as PDB; it can't access any userspace other than
the current process's and it can't make any changes to a running
process.  On the other hand, it's probably slightly less fragile.
The code's a lot cleaner, as well.