bitkeeper revision 1.1159.212.23 (41ee5e8fnbA6nB8wOl_As73BmvhklQ)
Support for USB virtualisation.  Adds the USB frontend / backend, update build options,
make the control tools understand USB interfaces.

One extra file from vanilla Linux is imported (linux-2.4.28-xen-sparse/drivers/usb/hcd.c)
because small changes were required to make XenU domains build with a USB frontend.  This
will not be required for the revised host controller infrastructure in Linux 2.6.

To use this, set "usb = " a list of USB ports in the domain config file.  Ports are
specified as a path, e.g. '1' = the first USB port on the host, '1/2' is the second port
on the hub on the first host port, etc.

Should work for most mass storage and isochronous devices.  May work for some HID devices
but these are not as intelligent and may have trouble supporting virtualisation.