bitkeeper revision 1.1236.1.62 (422996dcaXWGIhR3ZzSlNz3XpvLF0A)
[PATCH] [PATCH] gcc4isms part 2

This completes the gcc4ism changes to the Xen tree.  All of
the changes in this patch are in the tools/ subdirectory and
are pretty simple.  Some of the changes are for overly strict
gcc4 warnings, but I guess it's still good to verify that the
variable really wasn't unused, or that the sign change is
really harmless.

I guess I really do like -Werror now...

Signed-off-by: Rik van Riel <riel@redhat.com>

 libxc/xc_linux_build.c      |    2 +-
 libxc/xc_plan9_build.c      |   24 ++++++++++++------------
 libxc/xc_vmx_build.c        |    2 +-
 python/xen/lowlevel/xu/xu.c |    2 +-
 xcs/xcs.c                   |    2 +-
 xfrd/lzi_stream.c           |    2 +-
 xfrd/marshal.c              |    4 ++--
 xfrd/xdr.c                  |    2 +-
 xfrd/xfrd.c                 |   12 ++++++------
 9 files changed, 26 insertions(+), 26 deletions(-)