bitkeeper revision 1.1501 (428f097bke5QSeqibAc2HvxE7ue3aA)
[PATCH] pci-ioapic-x86-64.patch

Enable PCI, IOAPIC and ACPI on domain 0 for x86-64.

This patch is based on apic.c,io_apic.c and mpparse.c on x86-64. It
applies changes analogous to the i386 files for paravirtualization.

The patch was tested on UP and DP with acpi=on and off. We did this
on a 5/10 bk tree + Jun's fixes. The latest bk tree doesn't boot yet
and needs more work.

Signed-off-by: Arun Sharma <arun.sharma@intel.com>