bitkeeper revision 1.1730 (42ba7dbdkTADQfBqHaC_Ynqo3i1ZVw)
The xc_domain_getinfo does not clear the info structure before filling 
it in, and in some circumstances it doesn't completely fill in the 
structure for each domain.  In particular, it sets the crashed flag if 
the domain has crashed, but does not clear it if the domain has not 
crashed, so the crashed flag may be random stack garbage from the 
calling program.  With the attached patch (against the latest nightly), 
xc_domain_getinfo zeroes the array of info structures before filling 
them in.

Signed-off-by: Josh Triplett <josht@us.ibm.com>