Rename all dying domains to be prefixed with Zombie.  This allows a new domain to be created with the same name, fixing the race condition inside XendDomain that caused bug #278.
Move the state_set(TERMINATED) call onto the end of cleanupDomain rather than
destroyDomain, so that this flag is set when XendDomain cleans up a domain
that was killed without going through Xend.

Remove is_terminated, as this check is no longer necessary, since we are using
Zombie prefixes instead.

Signed-off-by: Ewan Mellor <ewan@xensource.com>