Remove all unnecessary uses of keyword arguments, to reduce complexity.  In many places we were supporting keywords for calling the methods here, and never using that feature (especially where all the arguments were compulsory anyway) and in some places we were even accepting positional and keyword arguments, and not actually taking any arguments to the method at all!
Fix the permission parsing string in xspy_set_permissions -- this would have
blown up horribly if anyone were actually calling this code.

Fix the documentation comments to include the transaction handle throughout.

Have the self parameter to all the methods be an XsHandle, not a PyObject, to
avoid casting.

Remove unused pyvalue_int and pyvalue_str.

Signed-off-by: Ewan Mellor <ewan@xensource.com>