Fix a race condition for multi-thread qemu dma, where vmx linux guests show warning "dma interrupt lost" and dma becomes very slow.
root cause: In the time between set ide irq and set dma status, if guest
receive the irq and query the status, it will find the status is not
ready and therefore treat it as pseudo interrupt. Change the order of
set irq and set dma status fixes this issue.

Signed-off-by: Ke Yu <ke.yu@intel.com>