[IA64] enable xencons irq for pcdp described uarts
This patch enables interrupts for xencons on UARTs described via the
HCDP/PCDP tables.  This gets a little messy because the old HCDP
doesn't provide interrupt polarity/trigger info.  If the UART is on a
PCI device we can assume it's low/level, otherwise we have to resort to
more creative methods.  Since this old table is being phased out in
favor of the new version that does provide all the interrupt info, we
should have a fixed number of boxes out there that we have to detect (or
leave in polling mode).  I've added a function to detect and setup the
HP models based on the model string found in the ACPI XSDT OEM table ID.
I also added a xencons_poll boot option in case detection goes wrong or
the table provides the wrong info, we can force it to a working state.
Removed all of the "About to call..." printks in xensetup while I was
there.

Signed-off-by: Alex Williamson <alex.williamson@hp.com>