This patch adds new xm subcommands to support working with resource labels.  The new subcommands are 'xm resources', 'xm rmlabel', 'xm getlabel' and 'xm dry-run'.  In addition, the 'xm addlabel' subcommand now uses an updated syntax to support labeling both domains and resources.  See the xm man page for details on each subcommand.
Beyond the new subcommands, this patch allows users to immediately see
when security checks will fail by pushing some basic security checking
into the beginning of 'xm create' and 'xm block-attach'.  ACM security
attributes for block devices are added to XenStore in order to support
the final security enforcement, which will be performed in the kernel
and included in a separate patch.

Signed-off-by: Bryan D. Payne <bdpayne@us.ibm.com>
Signed-off-by: Reiner Sailer <sailer@us.ibm.com>