[IA64] INIT hadler for support coredumping feature
This patch is for supporting coredumping feature.
Some dump feature is necessary to save registers into memory.
So I implemented it and add a hook of calling dump function.
Because all cpu have to save registers,
the same handler is called by monarch and slave handler.

This patch is useful not only dump feature, but also debug xen.
(Because INIT hadler is always able to show register and call trace.)

Signed-off-by: Akio Takebe <takebe_akio@jp.fujitsu.com>
[Updated for linux-xen location]
Signed-off-by: Alex Williamson <alex.williamson@hp.com>