[NET] front: Fix features on resume when csum is off
When the netfront driver is resumed the features are renegotiated with
the backend.  However, I forgot take into account the status of the TX
checksum setting.  When TX checksum is disabled by the user, we cannot
enable SG or TSO since both require checksum offload.  This patch
makes
xennet check the checksum setting before renegotiating SG or TSO.

This bug was fixed thanks to a report from Anton Burtsev.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>