[POWERPC][XEN] Detect bad spurious interrupt condition and panic instead of hang
When handing off the MPIC from Xen to Dom0, which is the current yet
not permamnet design, the MPIC can cause the processor to assert an
external interrupt when none is available.  Rather then simply hang in
this condition we now panic so the user can see that there is indeed a
problem and identify it as this one.

This condition seems to be related to temperature and the probablity
of it occuring decreases if the machine is allowed to stay idle (not
in the Xen panic loop) for a minute or two.

Signed-off-by: Jimi Xenidis <jimix@watson.ibm.com>
Signed-off-by: Hollis Blanchard <hollisb@us.ibm.com>