Small fix for xenoprof.
Update list of privileged xenoprof operations which were not in sync
with new added functions.
Since the number of privileged operations is larger than the non
privileged ones, change the logic to list unprivileged ones instead.
Also added a few printk statements on failed xenoprof initialization
to identify the source of the problem.

Signed-off-by: Jose Renato Santos <jsantos@hpl.hp.com>