[TPM] Add tests cases for the ACM security hooks and tools to the xm test suite. The tests can be run with ACM turned off (not compiled into Xen; see user doc for this), but most of them will be skipped then. They can be run with a command like
make -C tests/security-acm check-TESTS

from the xm-test directory. They are also part of the default tests in
the tests suite and part of a new group test 'security'.
Since some of the tests require resources to be labeled, one must
explicitly enable the resources to be allowed to be labeled by the
test
suite by running

./configure --enable-full-labeling

once. To turn it off, the '--enable-full-labeling' parameter should be
omitted.

Signed-off-by: Stefan Berger <stefanb@us.ibm.com>