tmem: (re-)enable by default
Late in the 4.0 release it was discovered that certain order>0
allocations could fail and had no fallback.  This conflicted with
tmem especially when combined with aggressive ballooning.
A hack-y workaround patch was added in time for 4.0 that has
reduced (but not completely eliminated) the problem but
tmem was left disabled-by-default for the 4.0 release.

Re-enable it in xen-unstable by default to help identify cases
where the workaround is insufficient.   Tmem can be
disabled with the no-tmem Xen boot option.  Please report
failures (that are fixed with the no-tmem option) to me.

Signed-off-by: Dan Magenheimer <dan.magenheimer@oracle.com>