xenpaging: optimize p2m_mem_paging_populate
p2m_mem_paging_populate will always put another request in the ring.
To reduce pressure on the ring, place only required requests in the
ring.  If the gfn was already processed by another thread, and the
current vcpu does not need to be paused, p2m_mem_paging_resume will do
nothing with the request.  And also xenpaging will drop the request if
the vcpu does not need a wakeup.

Signed-off-by: Olaf Hering <olaf@aepfle.de>