libxl: disks: commit libxlu_disk_l.[ch] flex output
Commit the flex output into revision control, for the benefit of those
without flex or with only very old flex.  This change is split into a
separate patch for ease of review.  Generated with flex 2.5.35-6 on
Debian lenny i386, by running "make" in the libxl directory.  No
manual changes in this commit.

Signed-off-by: Ian Jackson <Ian.Jackson@eu.citrix.com>
Committed-by: Ian Jackson <Ian.Jackson@eu.citrix.com>